Parser: Detect too short bin. single-q. string.
Previously, a too short single-quoted string would failed with the
`StringError::InvalidOpeningCharacter` error code, which was incorrect.
Now it fails with the `StringError::TooShort` error code, which is the
expected one.

Also, an out of range error is avoided by checking the length in the
middle of the binary string matching.